alloc is the new allocator
